fir for "The lat/lon fields have '-Inf' instead of fill values  (at least that's how they appears in IDL) where an in situ history was used for the in situ record; match-ups with a single in situ value have correct lat/lon fill values."

+++ src/main/java/org/esa/cci/sst/rules/TimeType.java
@@ -30,6 +30,6 @@ final class TimeType extends AbstractAttributeModification {
 
     @Override
     protected void configureTargetColumn(ColumnBuilder targetColumnBuilder, Item sourceColumn) {
-        targetColumnBuilder.type(DataType.INT).unit(Constants.UNIT_TIME).fillValue(Integer.MIN_VALUE);
+        targetColumnBuilder.type(DataType.INT).unit(Constants.UNIT_TIME).fillValue(Short.MIN_VALUE);
     }
 }
